fix for #691
git-svn-id: https://svn.kvirc.de/svn/trunk/kvirc@3895 17fca916-40b9-46aa-a4ea-0a15b648b75c

+++ src/kvilib/core/kvi_qstring.cpp
@@ -223,7 +223,7 @@ namespace KviQString
 		return (pC1 == pC1e);
 	}
 
-	QString makeSizeReadable(size_t bytes)
+	QString makeSizeReadable(quint64 bytes)
 	{
 		double dSize = bytes;
 		if(dSize < 1024)
